+++ substra/sdk/backends/base.py
@@ -29,6 +29,10 @@ class BaseBackend(abc.ABC):
     @abc.abstractmethod
     def update(self, key, spec, spec_options=None):
         raise NotImplementedError
+    
+    @abc.abstractmethod
+    def archive(self, key, spec, spec_options=None):
+        raise NotImplementedError
 
     @abc.abstractmethod
     def add_compute_plan_tuples(self, spec, spec_options):